requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --no-emit-index-url --output-file=requirements/tmp.txt subset.in
#
anyio==4.3.0
    # via
    #   -c requirements/requirements_dev.txt
    #   starlette
aramix-application-commons @ git+https://bitbucket.org/3rdplace/aramix-application-commons@v0.0.3
    # via
    #   -c requirements/requirements_dev.txt
    #   -r subset.in
aramix-security-commons @ git+https://bitbucket.org/3rdplace/aramix-security-commons@v0.0.1
    # via
    #   -c requirements/requirements_dev.txt
    #   hexagonal-core
cachetools==5.3.3
    # via
    #   -c requirements/requirements_dev.txt
    #   google-auth
certifi==2024.2.2
    # via
    #   -c requirements/requirements_dev.txt
    #   requests
cfg-load==0.9.0
    # via
    #   -c requirements/requirements_dev.txt
    #   py4ai-core
charset-normalizer==3.3.2
    # via
    #   -c requirements/requirements_dev.txt
    #   requests
click==8.1.7
    # via
    #   -c requirements/requirements_dev.txt
    #   uvicorn
deprecated==1.2.14
    # via
    #   -c requirements/requirements_dev.txt
    #   py4ai-core
exceptiongroup==1.2.0
    # via
    #   -c requirements/requirements_dev.txt
    #   anyio
fastapi==0.110.0
    # via
    #   -c requirements/requirements_dev.txt
    #   fastapi-utils
    #   microservices-core
    #   microservices-hexagonal
fastapi-utils==0.2.1
    # via
    #   -c requirements/requirements_dev.txt
    #   microservices-core
    #   microservices-hexagonal
google-api-core[grpc]==2.18.0
    # via
    #   -c requirements/requirements_dev.txt
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-firestore
google-auth==2.29.0
    # via
    #   -c requirements/requirements_dev.txt
    #   aramix-application-commons
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
google-cloud-bigquery==3.19.0
    # via
    #   -c requirements/requirements_dev.txt
    #   aramix-application-commons
google-cloud-bigquery-storage==2.24.0
    # via
    #   -c requirements/requirements_dev.txt
    #   aramix-application-commons
google-cloud-core==2.4.1
    # via
    #   -c requirements/requirements_dev.txt
    #   aramix-application-commons
    #   google-cloud-bigquery
    #   google-cloud-firestore
google-cloud-firestore==2.15.0
    # via
    #   -c requirements/requirements_dev.txt
    #   aramix-application-commons
google-crc32c==1.5.0
    # via
    #   -c requirements/requirements_dev.txt
    #   google-resumable-media
google-resumable-media==2.7.0
    # via
    #   -c requirements/requirements_dev.txt
    #   google-cloud-bigquery
googleapis-common-protos==1.63.0
    # via
    #   -c requirements/requirements_dev.txt
    #   google-api-core
    #   grpcio-status
grpcio==1.62.1
    # via
    #   -c requirements/requirements_dev.txt
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.1
    # via
    #   -c requirements/requirements_dev.txt
    #   google-api-core
h11==0.14.0
    # via
    #   -c requirements/requirements_dev.txt
    #   uvicorn
hexagonal-core @ git+https://bitbucket.org/3rdplace/hexagonal-core@v0.1.1
    # via
    #   -c requirements/requirements_dev.txt
    #   microservices-hexagonal
idna==3.6
    # via
    #   -c requirements/requirements_dev.txt
    #   anyio
    #   requests
microservices-core @ git+https://bitbucket.org/3rdplace/microservice-core@v0.0.1
    # via
    #   -c requirements/requirements_dev.txt
    #   microservices-hexagonal
microservices-hexagonal @ git+https://bitbucket.org/3rdplace/microservice-hexagonal@v0.1.1
    # via
    #   -c requirements/requirements_dev.txt
    #   -r subset.in
mpu[io]==0.23.1
    # via
    #   -c requirements/requirements_dev.txt
    #   cfg-load
numpy==1.26.4
    # via
    #   -c requirements/requirements_dev.txt
    #   pandas
    #   scipy
packaging==24.0
    # via
    #   -c requirements/requirements_dev.txt
    #   google-cloud-bigquery
pandas==2.2.1
    # via
    #   -c requirements/requirements_dev.txt
    #   py4ai-core
proto-plus==1.23.0
    # via
    #   -c requirements/requirements_dev.txt
    #   google-api-core
    #   google-cloud-bigquery-storage
    #   google-cloud-firestore
protobuf==4.25.3
    # via
    #   -c requirements/requirements_dev.txt
    #   google-api-core
    #   google-cloud-bigquery-storage
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
py4ai-core==1.0.2
    # via
    #   -c requirements/requirements_dev.txt
    #   aramix-application-commons
    #   aramix-security-commons
    #   hexagonal-core
    #   microservices-core
    #   microservices-hexagonal
pyasn1==0.5.1
    # via
    #   -c requirements/requirements_dev.txt
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via
    #   -c requirements/requirements_dev.txt
    #   google-auth
pydantic==1.10.14
    # via
    #   -c requirements/requirements_dev.txt
    #   aramix-application-commons
    #   aramix-security-commons
    #   fastapi
    #   fastapi-utils
    #   microservices-core
    #   microservices-hexagonal
    #   py4ai-core
python-dateutil==2.9.0.post0
    # via
    #   -c requirements/requirements_dev.txt
    #   google-cloud-bigquery
    #   pandas
pytz==2024.1
    # via
    #   -c requirements/requirements_dev.txt
    #   cfg-load
    #   mpu
    #   pandas
pyyaml==6.0.1
    # via
    #   -c requirements/requirements_dev.txt
    #   cfg-load
requests==2.31.0
    # via
    #   -c requirements/requirements_dev.txt
    #   cfg-load
    #   google-api-core
    #   google-cloud-bigquery
rsa==4.9
    # via
    #   -c requirements/requirements_dev.txt
    #   google-auth
scipy==1.12.0
    # via
    #   -c requirements/requirements_dev.txt
    #   py4ai-core
six==1.16.0
    # via
    #   -c requirements/requirements_dev.txt
    #   cfg-load
    #   python-dateutil
sniffio==1.3.1
    # via
    #   -c requirements/requirements_dev.txt
    #   anyio
sqlalchemy==1.4.52
    # via
    #   -c requirements/requirements_dev.txt
    #   fastapi-utils
starlette==0.36.3
    # via
    #   -c requirements/requirements_dev.txt
    #   fastapi
tomli==2.0.1
    # via
    #   -c requirements/requirements_dev.txt
    #   aramix-application-commons
    #   aramix-security-commons
    #   hexagonal-core
    #   microservices-core
    #   microservices-hexagonal
    #   py4ai-core
typing-extensions==4.10.0
    # via
    #   -c requirements/requirements_dev.txt
    #   anyio
    #   aramix-application-commons
    #   aramix-security-commons
    #   fastapi
    #   hexagonal-core
    #   microservices-core
    #   microservices-hexagonal
    #   py4ai-core
    #   pydantic
    #   uvicorn
tzdata==2024.1
    # via
    #   -c requirements/requirements_dev.txt
    #   pandas
tzlocal==5.2
    # via
    #   -c requirements/requirements_dev.txt
    #   mpu
urllib3==2.2.1
    # via
    #   -c requirements/requirements_dev.txt
    #   requests
uvicorn==0.29.0
    # via
    #   -c requirements/requirements_dev.txt
    #   microservices-core
    #   microservices-hexagonal
wrapt==1.16.0
    # via
    #   -c requirements/requirements_dev.txt
    #   deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
